fix: context.trigger_fraction is f64 so 0.85 round-trips config JSON exactly
An f32 0.85 serializes as 0.8500000238418579, which the settings UI showed
verbatim (the "0.00000000" report). f64 end-to-end: deacon ContextConfig,
agent CompressionConfig, threshold math.

+++ src/crates/regent-agent/src/application/lifecycle.rs
@@ -56,7 +56,7 @@ impl Agent {
         let estimate =
             compression::estimate_tokens(&self.system_prompt, self.transcript.messages());
         let threshold =
-            (self.config.max_context_tokens as f64 * f64::from(settings.trigger_fraction)) as u32;
+            (self.config.max_context_tokens as f64 * settings.trigger_fraction) as u32;
         if estimate <= threshold {
             return Ok(());
         }
